Move runtime handler to the cortext package
To avoid having a dependency of the validation package in the api one we
have move the runtime configuration handle to the cortex package.

This change also led to move some function of the handle to the util
package
