+++ pkg/cortex/runtime_config.go
@@ -3,10 +3,12 @@ package cortex
 import (
 	"errors"
 	"io"
+	"net/http"
 
 	"gopkg.in/yaml.v2"
 
 	"github.com/cortexproject/cortex/pkg/ring/kv"
+	"github.com/cortexproject/cortex/pkg/util"
 	"github.com/cortexproject/cortex/pkg/util/runtimeconfig"
 	"github.com/cortexproject/cortex/pkg/util/validation"
 )
@@ -81,5 +83,60 @@ func multiClientRuntimeConfigChannel(manager *runtimeconfig.Manager) func() <-ch
 		}()
 
 		return outCh
+	}
+}
+
+func diffLimitsConfig(defaultConfig, actualConfig map[interface{}]interface{}) (map[interface{}]interface{}, error) {
+	output := make(map[interface{}]interface{})
+	for tenant, tenantValues := range actualConfig {
+		tenantValuesObj, err := util.YAMLMarshalUnmarshal(tenantValues)
+		if err != nil {
+			return nil, err
+		}
+		tenantDiff, err := util.DiffConfig(defaultConfig, tenantValuesObj)
+		if err != nil {
+			return nil, err
+		}
+		output[tenant] = tenantDiff
+	}
+	return output, nil
+}
+
+func runtimeConfigHandler(runtimeCfgManager *runtimeconfig.Manager, defaultLimits validation.Limits) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var output interface{}
+		runtimeConfig := runtimeCfgManager.GetConfig()
+		if runtimeConfig == nil {
+			util.WriteTextResponse(w, "runtime config file doesn't exist")
+			return
+		}
+		switch r.URL.Query().Get("mode") {
+		case "diff":
+			defaultLimitsObj, err := util.YAMLMarshalUnmarshal(defaultLimits)
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			runtimeCfgObj, err := util.YAMLMarshalUnmarshal(runtimeConfig)
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			limitsCfgObj, err := util.YAMLMarshalUnmarshal(runtimeCfgObj["overrides"])
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			limitsDiff, err := diffLimitsConfig(defaultLimitsObj, limitsCfgObj)
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			runtimeCfgObj["overrides"] = limitsDiff
+			output = runtimeCfgObj
+		default:
+			output = runtimeConfig
+		}
+		util.WriteYAMLResponse(w, output)
 	}
 }
